feat: Implement Author Service and Controller with HTTP Methods and Feign Client for Book Retrieval - Implemented the AuthorService layer to handle business logic for author-related operations. - Created AuthorController with HTTP methods:   - POST `/api/authors`: Create a new author.   - GET `/api/authors/{id}`: Retrieve an author by ID.   - GET `/api/authors/{id}/books`: Retrieve an author with their associated books. - Integrated Feign Client for inter-service communication to fetch books for the author. - Debugged and resolved issues related to Feign Client and missing path in the request URL. - Updated `application.properties` to configure Eureka for service discovery and Feign Client.

+++ author_service/src/main/java/com/example/author_service/model/Author.java
@@ -13,7 +13,7 @@ public class Author {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "author_id")
-    private int authorId;
+    private Long authorId;
 
     @Column(name = "author_name", nullable=false)
     private String authorName;
@@ -30,7 +30,7 @@ public class Author {
     public Author() {
     }
 
-    public Author(int authorId, String authorName, String authorBiography, String nationality, String awards) {
+    public Author(Long authorId, String authorName, String authorBiography, String nationality, String awards) {
         this.authorId = authorId;
         this.authorName = authorName;
         this.authorBiography = authorBiography;
@@ -38,11 +38,11 @@ public class Author {
         this.awards = awards;
     }
 
-    public int getAuthorId() {
+    public Long getAuthorId() {
         return authorId;
     }
 
-    public void setAuthorId(int authorId) {
+    public void setAuthorId(Long authorId) {
         this.authorId = authorId;
     }
